+++ tests/scheduler/backends/test_qblox_backend.py
@@ -13,7 +13,6 @@ import json
 import logging
 import os
 import re
-import tempfile
 
 from typing import Dict, Generator
 
@@ -23,7 +22,6 @@ import pytest
 from pydantic import ValidationError
 from qblox_instruments import Pulsar, PulsarType
 
-from quantify_core.data.handling import set_datadir  # pylint: disable=no-name-in-module
 
 import quantify_scheduler
 from quantify_scheduler import Schedule
@@ -37,8 +35,8 @@ from quantify_scheduler.backends.qblox import (
 )
 from quantify_scheduler.backends.qblox.compiler_abc import Sequencer
 from quantify_scheduler.backends.qblox.helpers import (
-    convert_hw_config_to_portclock_configs_spec,
     assign_pulse_and_acq_info_to_devices,
+    convert_hw_config_to_portclock_configs_spec,
     generate_port_clock_to_device_map,
     find_all_port_clock_combinations,
     find_inner_dicts_containing_key,
@@ -68,7 +66,6 @@ from quantify_scheduler.compilation import (
     qcompile,
 )
 
-from quantify_scheduler.device_under_test.mock_setup import set_standard_params_transmon
 from quantify_scheduler.operations.acquisition_library import Trace
 from quantify_scheduler.operations.gate_library import Measure, Reset, X
 from quantify_scheduler.operations.pulse_library import (
@@ -857,8 +854,6 @@ def test_compile_simple(
     load_example_qblox_hardware_config,
 ):
     """Tests if compilation with only pulses finishes without exceptions"""
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
     qcompile(
         pulse_only_schedule,
         load_example_transmon_config,
@@ -871,8 +866,7 @@ def test_compile_cluster(
     cluster_only_schedule,
     load_example_qblox_hardware_config,
 ):
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
+
     qcompile(
         cluster_only_schedule,
         mock_setup_basic_transmon["quantum_device"].generate_device_config(),
@@ -881,8 +875,6 @@ def test_compile_cluster(
 
 
 def test_compile_no_device_cfg(load_example_qblox_hardware_config):
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
 
     sched = Schedule("One pulse schedule")
     sched.add_resources([ClockResource("q0.01", 3.1e9)])
@@ -902,8 +894,7 @@ def test_compile_simple_multiplexing(
     hardware_cfg_multiplexing,
 ):
     """Tests if compilation with only pulses finishes without exceptions"""
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
+
     qcompile(
         pulse_only_schedule_multiplexed,
         load_example_transmon_config,
@@ -917,8 +908,6 @@ def test_compile_identical_pulses(
     load_example_qblox_hardware_config,
 ):
     """Tests if compilation with only pulses finishes without exceptions"""
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
 
     compiled_schedule = qcompile(
         identical_pulses_schedule,
@@ -935,8 +924,6 @@ def test_compile_measure(
     load_example_transmon_config,
     load_example_qblox_hardware_config,
 ):
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
 
     full_program = qcompile(
         duplicate_measure_schedule,
@@ -965,9 +952,6 @@ def test_compile_clock_operations(
     operation: Operation,
     instruction_to_check: str,
 ):
-    # mock_setup_basic_transmon should arrange this but is not working here
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
 
     sched = Schedule("shift_clock_phase_only")
     sched.add(operation)
@@ -990,19 +974,18 @@ def test_compile_clock_operations(
 
 
 def test_compile_cz_gate(
-    mock_setup_basic_transmon, hardware_cfg_two_qubit_gate, two_qubit_gate_schedule
+    mock_setup_basic_transmon_with_standard_params,
+    hardware_cfg_two_qubit_gate,
+    two_qubit_gate_schedule,
 ):
-    # mock_setup_basic_transmon should arrange this but is not working here
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
 
-    set_standard_params_transmon(mock_setup_basic_transmon)
+    mock_setup = mock_setup_basic_transmon_with_standard_params
 
-    edge_q2_q3 = mock_setup_basic_transmon["q2_q3"]
+    edge_q2_q3 = mock_setup["q2_q3"]
     edge_q2_q3.cz.q2_phase_correction(44)
     edge_q2_q3.cz.q3_phase_correction(63)
 
-    quantum_device = mock_setup_basic_transmon["quantum_device"]
+    quantum_device = mock_setup["quantum_device"]
     device_cfg = quantum_device.generate_device_config()
 
     compiled_sched = qcompile(
@@ -1036,8 +1019,7 @@ def test_compile_simple_with_acq(
     load_example_transmon_config,
     load_example_qblox_hardware_config,
 ):
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
+
     full_program = qcompile(
         mixed_schedule_with_acquisition,
         load_example_transmon_config,
@@ -1058,13 +1040,12 @@ def test_compile_simple_with_acq(
     [True, False],
 )
 def test_compile_acq_measurement_with_clock_phase_reset(
-    mock_setup_basic_transmon,
+    mock_setup_basic_transmon_with_standard_params,
     load_example_qblox_hardware_config,
     reset_clock_phase,
 ):
-    set_standard_params_transmon(mock_setup_basic_transmon)
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
+    mock_setup = mock_setup_basic_transmon_with_standard_params
+
     schedule = Schedule("Test schedule")
 
     q0, q1 = "q0", "q1"
@@ -1081,8 +1062,8 @@ def test_compile_acq_measurement_with_clock_phase_reset(
             label=f"Measurement {q0}{i}",
         )
 
-    mock_setup_basic_transmon["q0"].measure.reset_clock_phase(reset_clock_phase)
-    device_cfg = mock_setup_basic_transmon["quantum_device"].generate_device_config()
+    mock_setup["q0"].measure.reset_clock_phase(reset_clock_phase)
+    device_cfg = mock_setup["quantum_device"].generate_device_config()
 
     compiled_schedule = qcompile(
         schedule, device_cfg, load_example_qblox_hardware_config
@@ -1103,8 +1084,7 @@ def test_acquisitions_back_to_back(
     load_example_transmon_config,
     load_example_qblox_hardware_config,
 ):
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
+
     sched = copy.deepcopy(mixed_schedule_with_acquisition)
     meas_op = sched.add(Measure("q0"))
     # add another one too quickly
@@ -1121,8 +1101,7 @@ def test_compile_with_rel_time(
     load_example_transmon_config,
     load_example_qblox_hardware_config,
 ):
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
+
     full_program = qcompile(
         pulse_only_schedule_with_operation_timing,
         load_example_transmon_config,
@@ -1140,8 +1119,7 @@ def test_compile_with_repetitions(
     load_example_transmon_config,
     load_example_qblox_hardware_config,
 ):
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
+
     mixed_schedule_with_acquisition.repetitions = 10
 
     full_program = qcompile(
@@ -1184,8 +1162,7 @@ def test_qasm_hook(pulse_only_schedule, load_example_transmon_config):
         },
     }
     sched = pulse_only_schedule
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
+
     sched.repetitions = 11
 
     full_program = qcompile(sched, load_example_transmon_config, hw_config)
@@ -1217,8 +1194,6 @@ def test_real_mode_pulses(
     hardware_cfg_real_mode,
     instruction_generated_pulses_enabled,  # pylint: disable=unused-argument
 ):
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
 
     real_square_pulse_schedule.repetitions = 10
     full_program = qcompile(
@@ -1381,13 +1356,11 @@ def test_container_prepare(
 
 
 def test_determine_scope_mode_acquisition_sequencer(
-    mock_setup_basic_transmon, load_example_qblox_hardware_config
+    mock_setup_basic_transmon_with_standard_params, load_example_qblox_hardware_config
 ):
     # mock_setup_basic_transmon should arrange this but is not working here
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
 
-    set_standard_params_transmon(mock_setup_basic_transmon)
+    mock_setup = mock_setup_basic_transmon_with_standard_params
     sched = Schedule("determine_scope_mode_acquisition_sequencer")
     sched.add(Measure("q0"))
     sched.add(Trace(duration=100e-9, port="q0:res", clock="q0.multiplex"))
@@ -1396,7 +1369,7 @@ def test_determine_scope_mode_acquisition_sequencer(
     hardware_cfg = load_example_qblox_hardware_config
     sched = qcompile(
         sched,
-        mock_setup_basic_transmon["quantum_device"].generate_device_config(),
+        mock_setup["quantum_device"].generate_device_config(),
         hardware_cfg,
     )
 
@@ -1534,8 +1507,6 @@ def test_real_mode_container(
 def test_assign_frequencies_baseband(
     load_example_transmon_config, load_example_qblox_hardware_config
 ):
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
 
     sched = Schedule("two_gate_experiment")
     sched.add(X("q0"))
@@ -1583,9 +1554,6 @@ def test_assign_frequencies_baseband_downconverter(
     load_example_transmon_config,
     load_example_qblox_hardware_config,
 ):
-
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
 
     sched = Schedule("two_gate_experiment")
     sched.add(X("q0"))
@@ -1659,8 +1627,6 @@ def test_assign_frequencies_baseband_downconverter(
 def test_assign_frequencies_rf(
     mock_setup_basic_transmon, load_example_qblox_hardware_config
 ):
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
 
     sched = Schedule("two_gate_experiment")
     sched.add(X("q2"))
@@ -1717,8 +1683,6 @@ def test_assign_frequencies_rf_downconverter(
     mock_setup_basic_transmon,
     load_example_qblox_hardware_config,
 ):
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
 
     sched = Schedule("two_gate_experiment")
     sched.add(X("q2"))
@@ -1798,8 +1762,6 @@ def test_assign_frequencies_rf_downconverter(
 
 
 def test_markers(mock_setup_basic_transmon, load_example_qblox_hardware_config):
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
 
     # Test for baseband
     sched = Schedule("gate_experiment")
@@ -1911,8 +1873,7 @@ def assembly_valid(compiled_schedule, qcm0, qrm0):
 def test_acq_protocol_append_mode_valid_assembly_ssro(
     dummy_pulsars, load_example_transmon_config, load_example_qblox_hardware_config
 ):
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
+
     repetitions = 256
     ssro_sched = readout_calibration_sched("q0", [0, 1], repetitions=repetitions)
     compiled_ssro_sched = qcompile(
@@ -1951,8 +1912,7 @@ def test_acq_protocol_append_mode_valid_assembly_ssro(
 def test_acq_protocol_average_mode_valid_assembly_allxy(
     dummy_pulsars, load_example_transmon_config, load_example_qblox_hardware_config
 ):
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
+
     repetitions = 256
     sched = allxy_sched("q0", element_select_idx=np.arange(21), repetitions=repetitions)
     compiled_allxy_sched = qcompile(
@@ -1992,8 +1952,6 @@ def test_acq_protocol_average_mode_valid_assembly_allxy(
 def test_acq_declaration_dict_append_mode(
     load_example_transmon_config, load_example_qblox_hardware_config
 ):
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
 
     repetitions = 256
 
@@ -2015,8 +1973,6 @@ def test_acq_declaration_dict_append_mode(
 def test_acq_declaration_dict_bin_avg_mode(
     load_example_transmon_config, load_example_qblox_hardware_config
 ):
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
 
     allxy = allxy_sched("q0")
     compiled_allxy_sched = qcompile(
@@ -2140,13 +2096,11 @@ def test_convert_hw_config_to_portclock_configs_spec(
     assert migrated_config == expected_config
 
     # Test that hardware_compile is converting automatically
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
 
     sched = make_basic_multi_qubit_schedule(["q0", "q1"])
     sched = device_compile(sched, load_example_transmon_config)
     with pytest.warns(
-        DeprecationWarning,
+        FutureWarning,
         match=r"hardware config adheres to a specification that is deprecated",
     ):
         hardware_compile(sched, old_config)
@@ -2160,9 +2114,6 @@ def test_apply_latency_corrections_invalid_raises(
     Providing an invalid latency correction specification raises an exception
     when compiling.
     """
-    # mock_setup should arrange this but is not working here
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
 
     sched = Schedule("Single Gate Experiment on Two Qubits")
     sched.add(X("q0"))
@@ -2186,18 +2137,15 @@ def test_apply_latency_corrections_invalid_raises(
 
 
 def test_apply_latency_corrections_valid(
-    mock_setup_basic_transmon, hardware_cfg_latency_corrections
+    mock_setup_basic_transmon_with_standard_params, hardware_cfg_latency_corrections
 ):
     """
     This test function checks that:
     Latency correction is set for the correct portclock key
     by checking against the value set in QASM instructions.
     """
-    # mock_setup_basic_transmon should arrange this but is not working here
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
 
-    set_standard_params_transmon(mock_setup_basic_transmon)
+    mock_setup = mock_setup_basic_transmon_with_standard_params
     sched = Schedule("Single Gate Experiment on Two Qubits")
     sched.add(X("q0"))
     sched.add(
@@ -2210,7 +2158,7 @@ def test_apply_latency_corrections_valid(
     hardware_cfg = hardware_cfg_latency_corrections
     compiled_sched = qcompile(
         schedule=sched,
-        device_cfg=mock_setup_basic_transmon["quantum_device"].generate_device_config(),
+        device_cfg=mock_setup["quantum_device"].generate_device_config(),
         hardware_cfg=hardware_cfg,
     )
 
@@ -2244,9 +2192,6 @@ def test_apply_latency_corrections_warning(
     Checks if warning is raised for a latency correction
     that is not a multiple of 4ns
     """
-    # mock_setup_basic_transmon should arrange this but is not working here
-    tmp_dir = tempfile.TemporaryDirectory()
-    set_datadir(tmp_dir.name)
 
     sched = Schedule("Single Gate Experiment")
     sched.add(