+++ tests/scheduler/backends/zhinst/test_settings.py
@@ -10,7 +10,6 @@ from unittest.mock import ANY, call
 
 import numpy as np
 import pytest
-from quantify_core.data.handling import set_datadir
 from zhinst.qcodes import base
 
 from quantify_scheduler import waveforms
@@ -30,9 +29,9 @@ def make_ufhqa(mocker) -> base.ZIBaseInstrument:
     return instrument
 
 
-def test_zi_settings_equality(tmp_test_data_dir):
+def test_zi_settings_equality(load_example_transmon_config):
     # Arrange
-    set_datadir(tmp_test_data_dir)
+
     sched_kwargs = {
         "pulse_amps": np.linspace(0, 0.5, 11),
         "pulse_duration": 1e-6,
@@ -47,7 +46,7 @@ def test_zi_settings_equality(tmp_test_data_dir):
         "repetitions": 10,
     }
     sched = awg_staircase_sched(**sched_kwargs)
-    device_cfg = load_json_example_scheme("transmon_test_config.json")
+    device_cfg = load_example_transmon_config
     hw_cfg = load_json_example_scheme("zhinst_test_mapping.json")
 
     hw_cfg["devices"][1]["channel_0"]["modulation"]["interm_freq"] = 10e6